Use FlatStyle.System for Next and Cancel buttons
This change enables Windows to render the button animations (mouse-over/out,
enable/disable). Using the default FlatStyle.Standard bypasses this
feature.

I did this change in the constructor because doing it in the WinForms
designer caused a giant diff due to its incorrectly reindenting the
InitializeComponent() function. Apparently the designer does not respect
my .editorconfig settings (not committed).