Fixing issue when the parent is a UserControl #15
The WizardControl assumes that the parent is always a Form. See WizardControl.OnParentChanged line 548. This prevents the ConfigureWindowFrame() from running in case the parent is a UserControl. Since the Form and the UserControl doesn't have a common base class that defines the Load event - which is triggering the previously mentioned function - we have to hook to both of those separate Load events, which makes the solution less elegant, but also fixes a nullrefence exception.